Move boshCredHubTemplate to bosh service and use autowired for credhubservice bean
