+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/bosh/BoshCredHubConfig.groovy
@@ -13,8 +13,10 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.swisscom.cloud.sb.broker.services.credhub
+package com.swisscom.cloud.sb.broker.services.bosh
 
+import com.swisscom.cloud.sb.broker.services.credhub.CertificateConfig
+import com.swisscom.cloud.sb.broker.services.credhub.CredHubConfig
 import groovy.transform.CompileStatic
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Configuration
